added the repo pattern to all models
