+++ app/models/holdings/Repository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Model\Repo;
+namespace App\Model\Repository;
 include_once __DIR__ . "/../../core/templates/DbTemplate.php";
 include_once __DIR__ . "/../../models/holdings/Entity.php";
